restructure the familyList
