+++ src/components/FamilyList/FamilyTable/index.ts
@@ -1,0 +1,1 @@
+export * from './FamilyTable';
